vim: Remove ConqueTerm
I never use this thing. Considering that I work in tmux, I tend to go
the other direction: push stuff out to a full-powered shell, not bring
the shell into Vim. My Dispatch.vim setup is in the works.

This also commits some YouCompleteMe settings and remarks, because I'm
getting ready to remove that too...