vim: Finally letting Command-T go for Ctrl-P
Command-T still works well, and Wincent is updating it more actively
than the Ctrl-P author, but I'm just getting weary of extra compilation
steps for Vim plugins when setting up a new machine. I stayed with
Command-T for a long time because it's much faster than Ctrl-P's native
search, but when using The Silver Searcher to drive the latter, it's on
par or better.