shell: Use mdls for UTI in Quick Look script unless using stdin
Always using MIME was throwing away better information when available,
like for Markdown files.

For the record since it's a mess to track down, I use Brett Terpstra's
fork of Fletcher Penney's MultiMarkdown Quick Look generator, to get the
extras of MultiMarkdown (vs. QLMarkdown) plus ability to hook in custom
CSS, and also patch it with additional file extension registrations that
ought to be in the original:

https://github.com/ttscoff/MMD-QuickLook
https://github.com/elmimmo/MMD-QuickLook/commit/258761745db9c126772afdf6e4e9c63285823793