shell: Avoid manual PYTHONPATH setting
It made switching between Python 2 and 3 environments problematic.

The solution (on Mavericks) for getting hg-git working with Homebrew
mercurial (which uses system Python) was to get pip installed via
`/usr/bin/easy_install --prefix=/usr pip` and get it installing packages
to /Library/Python/2.7/site-packages which is on the path.