tmux: Improve feedback, add session loading mappings
Display what's happening when toggling monitoring and logging.

Both approaches to loading sessions have some shortcomings but they
ought to be useful enough, it's not important enough to me at this point
to write a bunch of sidecar bash.

Be slightly more consistent with short/long command names. Usually I
like long names in scripts and config but some of tmux's short names are
plenty expressive so I'm not getting religious.